Fixes #110 - starting skill levels are now capped at 3 and don't jump on deserialize

+++ src/components/helpers/skill_t.hpp
@@ -7,7 +7,7 @@
 struct skill_t {
 	skill_t() {}
 	skill_t(const int8_t lvl, const uint16_t &xp) : skill_level(lvl), experience_gained(xp) {}
-	int8_t skill_level = 0;
+	int16_t skill_level = 0;
 	uint16_t experience_gained = 0;
 
 	void to_xml(rltk::xml_node * c);
